Implemented an RTOS over the existing architecture. All functionality is the same!
