+++ lib/src/driveTask.h
@@ -1,0 +1,11 @@
+//
+// Created by matt on 4/17/23.
+//
+
+#ifndef CARPROJECT_DRIVETASK_H
+#define CARPROJECT_DRIVETASK_H
+
+#endif //CARPROJECT_DRIVETASK_H
+
+
+void driveTask(void *parameter);
